+++ SRC/common.h
@@ -135,7 +135,7 @@ WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN 
 #elif __MACH__
 #include <sys/malloc.h>
 #elif FREEBSD
-O// avoid requesting malloc.h
+// avoid requesting malloc.h
 #else
 #include <malloc.h>
 #endif