+++ SRC/mainSystem.h
@@ -199,7 +199,7 @@ int main(int argc, char * argv[]);
 void ProcessOOB(char* buffer);
 void ComputeWhy(char* buffer, int n);
 void FlipResponses();
-
+int CountWordsInBuckets(int& unused, unsigned int* depthcount, int limit);
 // Input processing
 void MainLoop();
 void FinishVolley(char* input,char* output,char* summary,int limit = outputsize);